energy_model: Add EnergyModel class
This class provides a model of systems with:

 - CPU capacity at different frequencies.
 - Power usage at different frequency.
 - Power usage in different idle states.

The model is aware of topologically shared resources (clusters),
topological dependencies for idle states (power domains) and frequency
domains.

The intended use case for this model is for testing energy-aware
scheduling in a platform-agnostic way.

Also provided in this commit are:

 - Some very simple tests for EnergyModel
 - Examples of EnergyModel for Juno r0, HiKey and Pixel devices
 - Changes to TestEnv to allow the provided EnergyModels to be automatically
   loaded

This new interface provides scope for a lot of refactoring (for example the
`nrg_model` data could be removed), but that is left for a later date.