tools/wltests: add initial support for testing a series of patches
This provides a minimal set of scripts to support building, flashing
and booting an Android device for each kernel commits of a provided list.

The WLTest suite lives under LISA_HOME/tools/wltests and it consists of
these modules:

- test_series: the main script which allows to build, flash and reboot an
               Android target for a given set of SHA1 of a specified
               kernel source tree

- android/*.sh: a set of script to support the generation of Android
                specific images

- build: a set of functions to build different targets with proper
         abstractions for toolchains provisioning and generation of
         kernel builds configuration file

- helpers: a set of supporting functions

These scripts works based on a platform specific set of definitions which
is provided by a following set of patches.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>